+++ changelog.md
@@ -1,3 +1,8 @@
+## 3.1.0
+
+ - Interoperation with [ECMAScript Observables](https://github.com/zenparsing/es-observable) added with two new methods: `.fromESObservable` and `.toESObservable`/`[Symbol.observable]` [#154](https://github.com/rpominov/kefir/pull/154)
+ - A bug in `.fromEvents` fixed [#159](https://github.com/rpominov/kefir/issues/159)
+
 ## 3.0.0 (26/09/2015)
 
  - All previously deprecated methods removed. Full list can be found in [deprecated API docs of v2.x](https://github.com/rpominov/kefir/blob/v2/deprecated-api-docs.md).